chore: renames debuglogger into debuglog package.
