+++ debuglog/nop_test.go
@@ -1,7 +1,7 @@
 // Copyright 2022 Juan Pablo Tosso and the OWASP Coraza contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package debuglogger
+package debuglog
 
 import (
 	"io"
